refactor: 패키지 및 이름 변경 리팩토링
