+++ src/main/java/com/ssabae/nextstep/lotto/LottoMain.java
@@ -1,0 +1,15 @@
+package com.ssabae.nextstep.lotto;
+
+import com.ssabae.nextstep.lotto.application.LottoApplication;
+
+/**
+ * @author : leesangbae
+ * @project : java-lotto
+ * @since : 2020-12-19
+ */
+public class LottoMain {
+
+    public static void main(String[] args) {
+        new LottoApplication().launch();
+    }
+}
