+++ src/test/java/com/ssabae/nextstep/lotto/domain/lotto/ticketgenerator/AutoLottoTicketGeneratorTest.java
@@ -1,7 +1,10 @@
-package com.ssabae.nextstep.lotto.domain.lotto;
+package com.ssabae.nextstep.lotto.domain.lotto.ticketgenerator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.ssabae.nextstep.lotto.domain.lotto.LottoTicket;
+import com.ssabae.nextstep.lotto.domain.lotto.ticketgenerator.AutoLottoTicketGenerator;
+import com.ssabae.nextstep.lotto.domain.lotto.ticketgenerator.LottoTicketGenerator;
 import org.junit.jupiter.api.Test;
 
 /**
