+++ src/test/java/com/ssabae/nextstep/lotto/domain/lotto/ticketgenerator/TestLottoTicketGenerator.java
@@ -1,4 +1,7 @@
-package com.ssabae.nextstep.lotto.domain.lotto;
+package com.ssabae.nextstep.lotto.domain.lotto.ticketgenerator;
+
+import com.ssabae.nextstep.lotto.domain.lotto.LottoTicket;
+import com.ssabae.nextstep.lotto.domain.lotto.ticketgenerator.LottoTicketGenerator;
 
 /**
  * @author : leesangbae
